fix task 7

+++ 0x00-python_variable_annotations/7-to_kv.py
@@ -6,7 +6,7 @@
 from typing import Union, Tuple
 
 
-def to_kv(k: str, v: Union[int, float]) -> Tuple[str, int]:
+def to_kv(k: str, v: Union[int, float]) -> Tuple[str, float]:
     """_summary_
 
     Args:
